backend/requirements.txt
fastapi>=0.111.0
pydantic>=2.0.0
starlette>=0.37.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
requests>=2.31.0
python-dotenv>=1.0.1
Pillow>=10.3.0
aiofiles>=23.2.1
boto3>=1.34.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# AWS Packages
boto3>=1.34.0
botocore>=1.33.0
s3transfer>=0.7.0

# Azure Packages
azure-storage-blob>=12.19.0
azure-ai-textanalytics>=5.4.0
azure-core>=1.29.0
azure-signalr>=0.10.0

# Email & Text Processing
jinja2>=3.1.0
markdown>=3.5.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Additional Utilities
python-dateutil>=2.8.2

